feat: 판매 내역 조회 및 운영자용 기간별 정산 집계 API 구현

+++ creator-settlement-pjt/src/main/java/com/seohyun/creator_settlement_pjt/entity/SaleRecord.java
@@ -36,7 +36,7 @@ public class SaleRecord {
     private User student;
 
     @Column(name="paid_amount", nullable = false)
-    private int paidAmount;
+    private long paidAmount;
 
     @Column(name="paid_at", nullable = false)
     private LocalDateTime paidAt;
@@ -48,7 +48,7 @@ public class SaleRecord {
     private CancelRecord cancelRecord;
 
     @Builder
-    private SaleRecord(Course course, User student, int paidAmount, 
+    private SaleRecord(Course course, User student, long paidAmount, 
                         LocalDateTime paidAt, BigDecimal feeRate) {
         this.course = course;
         this.student = student;
